classe: ClienteHttp
