+++ Alura-stickers/src/App.java
@@ -1,10 +1,5 @@
 import java.io.InputStream;
-import java.net.URI;
 import java.net.URL;
-import java.net.http.HttpClient;
-import java.net.http.HttpRequest;
-import java.net.http.HttpResponse;
-import java.net.http.HttpResponse.BodyHandlers;
 import java.util.List;
 import java.util.Map;
 
@@ -14,15 +9,13 @@ public class App {
         // fazer uma conexão HTTP e buscar os top 250 filmes
         // String url = "https://api.mocki.io/v2/549a5d8b";
         String url = "https://api.mocki.io/v2/549a5d8b/NASA-APOD";
-        URI endereco = URI.create(url);
-        var client = HttpClient.newHttpClient();
-        var request = HttpRequest.newBuilder(endereco).GET().build();
-        HttpResponse<String> response = client.send(request, BodyHandlers.ofString());
-        String body = response.body();
+
+        var http = new ClienteHttp();
+        String json = http.buscaDados(url);
 
         // pegar só os dados que interessam (titulo, poster, classificação) [Parsear o json]
         var parser = new JsonParser();
-        List<Map<String, String>> listaDeConteudos = parser.parse(body);
+        List<Map<String, String>> listaDeConteudos = parser.parse(json);
 
         // exibir e manipular os dados
